Playground/MQTT-Smart-Object-Design/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "iot_mqtt_project"
version = "0.1"
dependencies = [
    "paho-mqtt==1.6.1",
    "gpxpy"
]

[tool.setuptools]
packages = { find = {} }